Rename multiplayer client classes
