+++ osu.Game/Tests/Visual/Multiplayer/MultiplayerTestScene.cs
@@ -19,7 +19,7 @@ namespace osu.Game.Tests.Visual.Multiplayer
         public const int PLAYER_1_ID = 55;
         public const int PLAYER_2_ID = 56;
 
-        [Cached(typeof(StatefulMultiplayerClient))]
+        [Cached(typeof(MultiplayerClient))]
         public TestMultiplayerClient Client { get; }
 
         [Cached(typeof(IRoomManager))]
